refactor: move formula helpers to sdk
